% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{get_game_info}
\alias{get_game_info}
\title{Game Info}
\usage{
get_game_info(game_id = NA, game_json = NA)
}
\arguments{
\item{game_id}{A character on numeric unique game id found in url of nwhl.zone game pages Defaults to NA}

\item{team_json}{If user wants to use pre-specified data frame from fromJSON function}
}
\description{
This function returns data on the game meta data json
}
\examples{
get_roster_info(18507502)
}
\keyword{game}
