% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{get_id_date}
\alias{get_id_date}
\title{Date ID Scraper}
\usage{
get_id_date(Season = NA, Year = NA, Month = NA, Day = NA)
}
\arguments{
\item{Season}{takes in the season as year1year2 (yyyyyyyy)}

\item{Year}{takes in a specific year formatted as yyyy}

\item{Month}{takes in a specifc month formatted as mm}

\item{Day}{takes in a specific day formatted as dd}
}
\description{
This function returns a vector of game ids given a specific date
}
\keyword{date}
