% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{get_id_schedule}
\alias{get_id_schedule}
\title{Schedule ID Scraper}
\usage{
get_id_schedule(Season = "20182019", teams = NA)
}
\arguments{
\item{Season}{takes in the season as year1year2 (yyyyyyyy), Defaults to 20182019}

\item{teams}{takes a specific team abbreviation, such as "BUF". Defaults to all teams}
}
\description{
This function returns a vector of game ids given a prespecified team and season
}
\examples{
get_id_schedule("20182019", "BUF")
}
\keyword{schedule}
