% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{get_play_by_play}
\alias{get_play_by_play}
\title{Play by Play Scraper}
\usage{
get_play_by_play(game_ids)
}
\arguments{
\item{game_ids}{character or numeric vector of unique game ids}
}
\description{
This function returns a data frame of the play by play in a tidy format
}
\examples{
get_play_by_play(c("22207010","22207013"))
}
\keyword{by}
\keyword{play}
