% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{get_player_summary}
\alias{get_player_summary}
\title{Player Summary Function}
\usage{
get_player_summary(pbp_df)
}
\arguments{
\item{pbp_df}{a dataframe consisting of play by play data from get_play_by_play function}
}
\description{
Given a tidy play by play file from get_play_by_play function, summarizes player stats at a game level
}
\keyword{player}
