% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{get_roster_info}
\alias{get_roster_info}
\title{Roster Info}
\usage{
get_roster_info(game_id = NA, roster_json = NA)
}
\arguments{
\item{game_id}{A character on numeric unique game id found in url of nwhl.zone game pages Defaults to NA}

\item{roster_json}{If user wants to use pre-specified data frame from fromJSON function}
}
\description{
This function retrieves a team's roster given a game id or the dataframe from the rjson output
}
\examples{
get_roster_info(18507502)
}
\keyword{roster}
