% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{get_team_info}
\alias{get_team_info}
\title{Team Info}
\usage{
get_team_info(game_id = NA, team_json = NA)
}
\arguments{
\item{game_id}{A character on numeric unique game id found in url of nwhl.zone game pages Defaults to NA}

\item{team_json}{If user wants to use pre-specified data frame from fromJSON function}
}
\description{
This function returns data on the team json data
}
\examples{
get_roster_info(18507502)
}
\keyword{team}
